Less destructive way of updating control version
Previously, `New-PCFControlVersion` would increase any found occurences of the build number where the build number was follow by a quote symbol. 

Test 1
Assuming an initial build number of 1 and a resource with an `order` of 1. Upon running `New-PCFControlVersion`, the control build number would be increased to 2, as would the resources order. This is because the resources order is followed by a quote.

Test 2
- Start a new PCF project
- Run `New-PCFControlVersion` 7 time on the manifest file
- On the 8th run, note the `encoding` attribute of the `xml` node it changed to 'UTF-9', rendering the file invalid

The proposed way uses the `xml` and `version` types for a more readible and reliable approach.